Arquillian Weld Containers requires 'ArchiveAsset's to scan for nested classes
The Weld Container attempts to fake a WebArchive and support nested jars by
scanning for *.class and nested ArchiveAssets. ArchiveAssets are ShrinkWrap'ed
Archives and can be scanned for resources.

When adding a jar as a File, ShrinkWrap will add it as a FileAsset and the
Weld Container is not able to scan the content of it.

In this case, due to leaking from the ApplicationClassLoader, all the DeltaSpike
extensions META-INF/services and META-INF/beans.xml are found, but only the
Classes found the @Deployment are registered as Beans.

A DeltaSpike extension is then activated due to the ApplicationClassloader leak
and is expecting to find one of it's own Beans in the BeanManager. But since the
nested Jars are never scanned it's never registered as a Bean.